pyproject.toml
[project]
name = "excelifai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "groq>=0.30.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "pymupdf>=1.26.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
]
